refactor(): correct directory name
